chore(tests): add pytest markers and switch to marker-based test filtering
Add pytestmark = pytest.mark.unit to all unit test files and
pytestmark = pytest.mark.e2e to the missing e2e test file. Replace
the conftest collection hook with -m unit/-m e2e in the Makefile.
Move --e2e-debug logging setup to tests/e2e/conftest.py.
