+++ .pre-commit-config.yaml
@@ -49,7 +49,7 @@ repos:
     hooks:
       - id: pytest
         name: pytest
-        entry: uv run --extra dev pytest
+        entry: uv run --extra dev pytest -m unit
         language: system
         pass_filenames: false
         always_run: true
