+++ Makefile
@@ -4,12 +4,12 @@ pre-commit:
 
 .PHONY: test
 test:
-	uv run pytest --color=yes -ra
+	uv run pytest -m unit --color=yes -ra
 
 .PHONY: test-e2e
 test-e2e:
 	uv run uv pip show eval-hub-server
-	uv run pytest --e2e --e2e-debug -s -x --color=yes -ra
+	uv run pytest -m e2e --e2e-debug -s -x --color=yes -ra
 
 .PHONY: ruff
 ruff:
